backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Om Joshi <39465356+backpack455@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
weaviate-client = "^4.8.1"
langchain = "^0.3.3"
langchain-community = "^0.3.2"
langchain-weaviate = "^0.0.3"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
